+++ packages/react-combobox/src/components/Combobox/useComboboxStyles.ts
@@ -1,7 +1,14 @@
+import { SlotClassNames } from '@fluentui/react-utilities';
 import { makeStyles, mergeClasses } from '@griffel/react';
-import type { ComboboxState } from './Combobox.types';
+import type { ComboboxSlots, ComboboxState } from './Combobox.types';
 
+/**
+ * @deprecated Use `comboboxClassNames.root` instead.
+ */
 export const comboboxClassName = 'fui-Combobox';
+export const comboboxClassNames: SlotClassNames<ComboboxSlots> = {
+  root: 'fui-Combobox',
+};
 
 /**
  * Styles for the root slot
@@ -15,7 +22,7 @@ const useStyles = makeStyles({
  */
 export const useComboboxStyles_unstable = (state: ComboboxState): ComboboxState => {
   const styles = useStyles();
-  state.root.className = mergeClasses(comboboxClassName, styles.root, state.root.className);
+  state.root.className = mergeClasses(comboboxClassNames.root, styles.root, state.root.className);
 
   return state;
 };