+++ packages/react-provider/src/components/FluentProvider/useFluentProviderStyles.ts
@@ -1,8 +1,15 @@
 import { makeStyles, mergeClasses } from '@griffel/react';
 import { tokens } from '@fluentui/react-theme';
-import type { FluentProviderState } from './FluentProvider.types';
+import type { FluentProviderSlots, FluentProviderState } from './FluentProvider.types';
+import { SlotClassNames } from '@fluentui/react-utilities';
 
+/**
+ * @deprecated Use `fluentProviderClassNames.root` instead.
+ */
 export const fluentProviderClassName = 'fui-FluentProvider';
+export const fluentProviderClassNames: SlotClassNames<FluentProviderSlots> = {
+  root: 'fui-FluentProvider',
+};
 
 const useStyles = makeStyles({
   root: {
@@ -18,7 +25,12 @@ const useStyles = makeStyles({
 export const useFluentProviderStyles_unstable = (state: FluentProviderState) => {
   const styles = useStyles();
 
-  state.root.className = mergeClasses(fluentProviderClassName, state.themeClassName, styles.root, state.root.className);
+  state.root.className = mergeClasses(
+    fluentProviderClassNames.root,
+    state.themeClassName,
+    styles.root,
+    state.root.className,
+  );
 
   return state;
 };