+++ packages/theme/CHANGELOG.md
@@ -1,8 +1,27 @@
 # Change Log - @fluentui/theme
 
-This log was last generated on Thu, 10 Mar 2022 07:34:50 GMT and should not be manually modified.
+This log was last generated on Tue, 15 Mar 2022 07:45:54 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [2.6.4](https://github.com/microsoft/fluentui/tree/@fluentui/theme_v2.6.4)
+
+Tue, 15 Mar 2022 07:45:54 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/theme_v2.6.3..@fluentui/theme_v2.6.4)
+
+### Patches
+
+- Bump @fluentui/utilities to v8.8.1 ([PR #22094](https://github.com/microsoft/fluentui/pull/22094) by beachball)
+
+## [2.6.3](https://github.com/microsoft/fluentui/tree/@fluentui/theme_v2.6.3)
+
+Fri, 11 Mar 2022 07:34:41 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/theme_v2.6.2..@fluentui/theme_v2.6.3)
+
+### Patches
+
+- Bump @fluentui/merge-styles to v8.5.0 ([PR #22047](https://github.com/microsoft/fluentui/pull/22047) by beachball)
+- Bump @fluentui/utilities to v8.8.0 ([PR #22047](https://github.com/microsoft/fluentui/pull/22047) by beachball)
 
 ## [2.6.2](https://github.com/microsoft/fluentui/tree/@fluentui/theme_v2.6.2)